sampleのファイル差し替え、svgobjectgeneratorを一つにする
