+++ Sources/FileGenKit/SVG/SVGObjectGenerator.swift
@@ -2,14 +2,14 @@ import Data
 import Cocoa
 
 class SVGObjectGenerator {
-    private let views: [View]
+    private let columnViewsList: [ColumnViews]
     private let settings: Settings
     
     var svgObjectList: [SVGObjectType] {  _svgObjectList }
     private var _svgObjectList: [SVGObjectType] = []
     
-    init(views: [View], settings: Settings) {
-        self.views = views
+    init(columnViewsList: [ColumnViews], settings: Settings) {
+        self.columnViewsList = columnViewsList
         self.settings = settings
     }
     
@@ -19,122 +19,132 @@ class SVGObjectGenerator {
         let viewHorizontalMargin: Double = settings.viewHorizontalMargin
         let viewObjectSize: NSSize = settings.viewObjectSize
         
-        views.enumerated().forEach { data in
-            // 第一階層
-            let view: View = data.element
-            let index: Int = data.offset
-            
-            let preViewMaxCount: Int = {
-                if index != 0 {
-                    let views = views[0...(index - 1)].map { $0 }
-                    // TODO: ここが浮いているのでもう少し改善したい。設置したViewListのうち最も高いものを算出すればできそうな気がする
-                    return FileHeightCalculator.calculateMaxVerticalCount(views: views)
-                } else {
-                    return 0
-                }
+        appendSVGObject()
+        appendSVGLineObject()
+        appendSVGRectObject()
+        
+        func appendSVGObject() {
+            let sideMargin: Double = margin * 2
+            let width: Double = {
+                let objectWidth: Double = Double(columnViewsList.count) * viewObjectSize.width
+                let objectBetweenMargin: Double = Double(columnViewsList.count - 1) * viewHorizontalMargin
+                return objectWidth + objectBetweenMargin + sideMargin
+            }()
+            let height: Double = {
+                let maxLineNumber = columnViewsList
+                    .reduce(into: [Int]()) { partialResult, columnViews in
+                        let maxLineNumber = columnViews.viewList
+                            .sorted(by: { $0.lineNumber > $1.lineNumber })
+                            .first?.lineNumber
+                        if let _maxLineNumber = maxLineNumber {
+                            partialResult.append(_maxLineNumber)
+                        }
+                    }
+                    .max()
+                
+                let lineNumber = Double(maxLineNumber ?? 0) + 1 // lineNumberは0から始まるので
+                let objectHeight: Double = lineNumber * viewObjectSize.height
+                let objectBetweenMargin: Double = (lineNumber - 1) * viewVerticalMargin
+                return objectHeight + objectBetweenMargin + sideMargin
             }()
             
-            let viewY = margin + (Double(preViewMaxCount) * (viewVerticalMargin + viewObjectSize.height))
-            generateRect(x: margin, y: viewY, view: view)
-            
-            // 第二階層以降
-            generateRectsFromViews(view.views, baseViewY: viewY)
-            
-            // 斜線
-            generateLinesFromViews(baseView: view)
-            
-            func generateRectsFromViews(_ views: [View], baseViewY: Double) {
-                var nextViews: [View] = []
-                views.enumerated().forEach { data in
-                    let _view: View = data.element
-                    let _viewOffSet: Int = data.offset
+            _svgObjectList.append(
+                .svg(width: width, height: height)
+            )
+        }
+        
+        func appendSVGRectObject() {
+            columnViewsList.forEach { columnViews in
+                columnViews.viewList.forEach { data in
+                    let view: View = data.view
+                    let lineNumber: Int = data.lineNumber
+                                    
+                    let x: Double = margin + (Double(columnViews.columnNumber) * (viewHorizontalMargin + viewObjectSize.width))
+                    let y: Double = margin + (Double(lineNumber) * (viewVerticalMargin + viewObjectSize.height))
                     
-                    let x: Double = (Double(_view.index) * (viewHorizontalMargin + viewObjectSize.width)) + margin
-                    let y: Double = Double(_viewOffSet) * (viewVerticalMargin + viewObjectSize.height) + baseViewY
+                    let fillColorStr: String = view.contentColor.isEmpty == false
+                     ? view.contentColor
+                     : settings.viewObjectColorStr
+                    let strokeColorStr: String = view.borderColor.isEmpty == false
+                     ? view.borderColor
+                     : settings.viewObjectBorderColorStr
                     
-                    generateRect(x: x, y: y, view: _view)
+                    let rectSvg: SVGObjectType = .rect(x: x,
+                                                       y: y,
+                                                       width: viewObjectSize.width,
+                                                       height: viewObjectSize.height,
+                                                       fill: fillColorStr,
+                                                       stroke: strokeColorStr)
+                    let textSvg: SVGObjectType = .text(x: x + 5,
+                                                       y: y + 20,
+                                                       fontSize: settings.viewObjectTextFontSize,
+                                                       fill: settings.viewObjectTextColorStr,
+                                                       value: view.nameData.key)
                     
-                    if _view.views.count > 0 {
-                        nextViews.append(contentsOf: _view.views)
-                    }
-                }
-                
-                if nextViews.isEmpty == false {
-                    generateRectsFromViews(nextViews, baseViewY: baseViewY)
-                }
-            }
-            
-            func generateRect(x: Double, y: Double, view: View) {
-                // view.rectを使用してlineのrectが算出されることに注意
-                view.updateRect(.init(x: x,
-                                      y: y,
-                                      width: viewObjectSize.width,
-                                      height: viewObjectSize.height))
-                
-                let fillColorStr: String = view.contentColor.isEmpty == false
-                 ? view.contentColor
-                 : settings.viewObjectColorStr
-                let strokeColorStr: String = view.borderColor.isEmpty == false
-                 ? view.borderColor
-                 : settings.viewObjectBorderColorStr
-                
-                let rectSvg: SVGObjectType = .rect(x: x,
-                                                   y: y,
-                                                   width: viewObjectSize.width,
-                                                   height: viewObjectSize.height,
-                                                   fill: fillColorStr,
-                                                   stroke: strokeColorStr)
-                let textSvg: SVGObjectType = .text(x: x + 5,
-                                                   y: y + 20,
-                                                   fontSize: settings.viewObjectTextFontSize,
-                                                   fill: settings.viewObjectTextColorStr,
-                                                   value: view.nameData.key)
-                
-                if view.urlStr.isEmpty == false {
-                    _svgObjectList.append(.url(urlStr: view.urlStr, rect: rectSvg, text: textSvg))
-                } else {
-                    _svgObjectList.append(rectSvg)
-                    _svgObjectList.append(textSvg)
-                }
-            }
-            
-            func generateLinesFromViews(baseView: View) {
-                baseView.views.enumerated().forEach { data in
-                    let _view: View = data.element
-                    let startPoint = NSPoint(x: baseView.rect.maxX, y: baseView.rect.minY + (settings.viewObjectSize.height / 2))
-                    let endPoint = NSPoint(x: _view.rect.minX, y: _view.rect.minY + (settings.viewObjectSize.height / 2))
-                    
-                    let lineColor: String = {
-                        let defaultTransitionTypeKey = settings.transitionTypeList
-                            .first { $0.isDefault }?.typeStr ?? settings.transitionTypeList.first?.typeStr
-                        let filteredtransitionTypeKey = _view.transitionTypeKey.isEmpty == false
-                                                             ? _view.transitionTypeKey
-                                                             : defaultTransitionTypeKey
-                        
-                        if let defaultContext = settings.transitionTypeList
-                            .first(where: { $0.typeStr == filteredtransitionTypeKey }) {
-                            return defaultContext.colorStr
-                        } else {
-                            return "000000"
-                        }
-                    }()
-                    
-                    _svgObjectList.append(
-                        .line(x1: startPoint.x,
-                              y1: startPoint.y,
-                              x2: endPoint.x,
-                              y2: endPoint.y,
-                              stroke: lineColor,
-                              strokeWidth: settings.lineWidth,
-                              isMarker: false)
-                    )
-                                
-                    if _view.views.count > 0 {
-                        generateLinesFromViews(baseView: _view)
+                    if view.urlStr.isEmpty == false {
+                        _svgObjectList.append(.url(urlStr: view.urlStr, rect: rectSvg, text: textSvg))
+                    } else {
+                        _svgObjectList.append(rectSvg)
+                        _svgObjectList.append(textSvg)
                     }
                 }
             }
         }
         
+        func appendSVGLineObject() {
+            for columnViews in columnViewsList {
+                let columnNumber: Double = Double(columnViews.columnNumber)
+                if columnNumber == 0 {
+                    continue
+                }
+                
+                columnViews.viewList.forEach { viewInfo in
+                    let y: Double = margin + ((viewObjectSize.height + viewVerticalMargin) * Double(viewInfo.lineNumber)) + (viewObjectSize.height / 2)
+                    let x2: Double = margin + ((viewObjectSize.width + viewHorizontalMargin) * columnNumber)
+                    
+                    let x1: Double = {
+                        if viewInfo.transitionData.number == 0 {
+                            return x2 - viewHorizontalMargin
+                        } else {
+                            return x2 - (viewHorizontalMargin / 2)
+                        }
+                    }()
+                    
+                    _svgObjectList.append(
+                        .line(x1: x1,
+                              y1: y,
+                              x2: x2,
+                              y2: y,
+                              stroke: "000000",
+                              strokeWidth: settings.lineWidth,
+                              isMarker: true)
+                    )
+                    
+                    // 縦線
+                    if viewInfo.transitionData.number == 0 {
+                        let key = viewInfo.transitionData.sourceViewKey
+                        let filterViewInfoList = columnViews.viewList
+                            .filter { $0.transitionData.sourceViewKey == key }
+                        
+                        if filterViewInfoList.count > 1,
+                           let endLineNumber = filterViewInfoList.sorted(by: { $0.lineNumber > $1.lineNumber }).first?.lineNumber {
+                               let x = x2 - (viewHorizontalMargin / 2)
+                               let y1: Double = y
+                               let y2: Double = margin + ((viewObjectSize.height + viewVerticalMargin) * Double(endLineNumber)) + (viewObjectSize.height / 2)
+                               
+                               _svgObjectList.append(
+                                .line(x1: x,
+                                      y1: y1,
+                                      x2: x,
+                                      y2: y2,
+                                      stroke: "000000",
+                                      strokeWidth: settings.lineWidth,
+                                      isMarker: false)
+                               )
+                           }
+                    }
+                }
+            }
+        }
     }
 }
